Refactor settings components and introduce BrowserSettings
- Removed conditional rendering for the settings button in the Header component.
- Added a new BrowserSettings component to manage browser-related settings.
- Updated the Settings component to include BrowserSettings and modified state management for current tab selection.
- Adjusted QuerySettings to accept a handleSubmit prop and changed its structure to a div instead of a form.
